+++ CHANGELOG.md
@@ -44,7 +44,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 - Removed the ``hsn-bringup`` stage of ``sat bootsys boot`` due to removal of
-  underlying Ansible playbook and old fabric controller service.
+  underlying Ansible playbook, ``ncmp_hsn_bringup.yaml``, and old fabric
+  controller service.
+- Removed call to ``ncmp_hsn_bringup.yaml`` Ansible playbook from currently
+  unused ``HSNBringupWaiter`` class.
 - Removed calls to removed Ansible playbooks ``enable-dns-conflict-hosts.yml``
   and ``disable-dns-conflict-hosts.yml`` during ``ncn-power`` stage of
   ``sat bootsys``.